move Heading to common folder
