--- src/components/Scroll/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Scroll';
