+++ src/components/common/Scroll/index.ts
@@ -1,0 +1,1 @@
+export * from './Scroll';
