+++ src/pages/MessagePage/MessagePage.tsx
@@ -1,12 +1,10 @@
 import React, { FC, useRef, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import cn from 'classnames';
-import { formatDateTime } from '../../utils';
 
-import { Button, NavButton } from '../../components';
-import Heading from '../../components/Heading';
+import { formatDateTime } from '../../utils';
+import { Button, NavButton, Scroll, Heading } from '../../components';
 import Message from '../../components/ForumPageComponents/Message';
-import Scroll from '../../components/Scroll';
 
 import { out_arr } from './mock';
 import { topic_arr } from './mock';
