+++ src/pages/SignupPage/SignupPage.tsx
@@ -1,7 +1,6 @@
 import React, { FC, FormEvent, useCallback } from 'react';
 
-import { Button, NavButton, Input, Form } from '../../components';
-import Heading from '../../components/Heading';
+import { Button, NavButton, Input, Form, Heading } from '../../components';
 
 import { MenuAction, MenuActionType } from '../../types';
